+++ src/main/webapp/components/containers/index.tsx
@@ -1,3 +1,3 @@
-const FilterCard = require('./filter-card')
+const FilterCard = require('./filter-card').default
 import SpacedLinearContainer from './spaced-linear-container'
 export { FilterCard, SpacedLinearContainer }
