+++ src/main/webapp/components/query-selector/index.js
@@ -1,0 +1,2 @@
+import QuerySelector from './query-selector'
+export { QuerySelector as default }
