+++ src/main/webapp/components/visualizations/index.js
@@ -1,0 +1,2 @@
+import Visualizations from './visualizations'
+export default Visualizations
