+++ src/main/webapp/components/workspaces/index.js
@@ -1,0 +1,3 @@
+import Workspaces, { Workspace } from './workspaces'
+
+export { Workspaces, Workspace }
